Update_2.0.0/Refactor: divide godlike file manager
 - Divide the godlike file manager.py into the
creation.py, joining.py and overview.py.
 - Fix warnings.
